API/DERMDETECT-API/requirements.txt
Flask
python-dotenv
tensorflow
numpy
pillow
